+++ src/main/java/vn/kms/ngaythobet/domain/statistic/PlayerStatisticService.java
@@ -11,8 +11,11 @@ import org.springframework.transaction.annotation.Transactional;
 import vn.kms.ngaythobet.domain.betting.BettingMatch;
 import vn.kms.ngaythobet.domain.betting.BettingMatchRepository;
 import vn.kms.ngaythobet.domain.betting.BettingPlayer;
+import vn.kms.ngaythobet.domain.betting.BettingPlayerRepository;
+import vn.kms.ngaythobet.domain.core.User;
 import vn.kms.ngaythobet.domain.tournament.Competitor;
 import vn.kms.ngaythobet.domain.tournament.Match;
+import vn.kms.ngaythobet.domain.util.DataInvalidException;
 import vn.kms.ngaythobet.domain.util.SecurityUtil;
 import vn.kms.ngaythobet.web.dto.PlayerStatisticInfo;
 
@@ -22,14 +25,16 @@ public class PlayerStatisticService {
 
     private final BettingMatchRepository bettingMatchRepo;
 
+    private final BettingPlayerRepository bettingPlayerRepo;
+
     @Autowired
-    public PlayerStatisticService(BettingMatchRepository bettingMatchRepo) {
+    public PlayerStatisticService(BettingMatchRepository bettingMatchRepo, BettingPlayerRepository bettingPlayerRepo) {
         this.bettingMatchRepo = bettingMatchRepo;
+        this.bettingPlayerRepo = bettingPlayerRepo;
     }
 
     public TotalPlayerStatistic playerStatistic(PlayerStatisticInfo playerStatisticInfo) {
         String username = SecurityUtil.getCurrentLogin();
-        StatisticUtils statisticUtils = new StatisticUtils();
         TotalPlayerStatistic totalPlayerStatistic = new TotalPlayerStatistic();
         List<PlayerStatistic> playerStatistics = new ArrayList<>();
         double totalLossAmount = 0;
@@ -50,20 +55,22 @@ public class PlayerStatisticService {
                     playerStatistic.setCompetitor1Score(match.getScore1());
                     playerStatistic.setCompetitor2Score(match.getScore2());
                 } else {
-                  //doesn't have score here but Long - datatype auto set value = 0 if it null -> set it = -1 and UI will check it and show right data to user
-                    playerStatistic.setCompetitor1Score(-1); 
-                    playerStatistic.setCompetitor2Score(-1); 
+                    // doesn't have score here but Long - datatype auto set
+                    // value = 0 if it null -> set it = -1 and UI will check it
+                    // and show right data to user
+                    playerStatistic.setCompetitor1Score(-1);
+                    playerStatistic.setCompetitor2Score(-1);
                 }
                 // count lost amount when user bet
                 if (bettingPlayer.isPresent()) {
                     Competitor betCompetitor = bettingPlayer.get().getBetCompetitor();
                     playerStatistic.setBetCompetitorName(betCompetitor.getName());
-                    playerStatistic.setLossAmount(statisticUtils.calculateLossAmount(bettingMatch, betCompetitor));
+                    playerStatistic.setLossAmount(StatisticUtils.calculateLossAmount(bettingMatch, betCompetitor));
                 }
                 // count lost amount when user didnot bet
                 else {
                     playerStatistic.setBetCompetitorName("--");
-                    playerStatistic.setLossAmount(statisticUtils.calculateLossAmount(bettingMatch, null));
+                    playerStatistic.setLossAmount(StatisticUtils.calculateLossAmount(bettingMatch, null));
                 }
                 playerStatistics.add(playerStatistic);
                 totalLossAmount += playerStatistic.getLossAmount();
@@ -73,5 +80,21 @@ public class PlayerStatisticService {
         totalPlayerStatistic.setPlayerStatistics(playerStatistics);
         totalPlayerStatistic.setTotalLossAmount(totalLossAmount);
         return totalPlayerStatistic;
+    }
+
+    public double getLostAmountByUser(Long bettingMatchId) {
+        User player = SecurityUtil.getCurrentLoginUser();
+        BettingMatch bettingMatch = bettingMatchRepo.findOne(bettingMatchId);
+        if (bettingMatch == null) {
+            throw new DataInvalidException("validation.existMatchEntity.message");
+        } else if (bettingMatch.getMatch().getScore1() == null || bettingMatch.getMatch().getScore2() == null) {
+            return 0;
+        } else {
+            BettingPlayer bettingPlayer = bettingPlayerRepo.findByPlayerAndBettingMatch(player, bettingMatch);
+            if (bettingPlayer != null) {
+                return StatisticUtils.calculateLossAmount(bettingMatch, bettingPlayer.getBetCompetitor());
+            }
+            return 0;
+        }
     }
 }