+++ src/main/webui/app/components/unAuthorized/unAuthorized.js
@@ -2,7 +2,8 @@
 
 export default class UnAuthorizedController {
   /* @ngInject */
-  constructor($timeout, $location) {
+  constructor($timeout, $location, $rootScope) {
+    $rootScope.$broadcast('authen');
     $timeout(function () {
       $location.path('/home');
     }, 3000);