+++ src/test/java/vn/kms/ngaythobet/domain/tournament/GroupServiceTest.java
@@ -4,6 +4,7 @@ import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
 
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 
@@ -14,6 +15,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import vn.kms.ngaythobet.BaseTest;
 import vn.kms.ngaythobet.domain.core.User;
+import vn.kms.ngaythobet.domain.core.User.Role;
 import vn.kms.ngaythobet.domain.util.DataInvalidException;
 import vn.kms.ngaythobet.web.dto.AddNewMemberInfo;
 import vn.kms.ngaythobet.web.dto.CheckModeratorInfo;
@@ -42,14 +44,10 @@ public class GroupServiceTest extends BaseTest {
         tournament.setName("World Cup");
         tournament.setNumOfCompetitor((long) 32);
         tournament.setGroups(Collections.emptyList());
-        Tournament temp = tournamentRepo.save(tournament);
+        tournamentRepo.save(tournament);
 
         user = getDefaultUser();
-        CreateGroupInfo createGroupInfo = new CreateGroupInfo();
-        createGroupInfo.setName("Launch 4");
-        createGroupInfo.setTournamentId(temp.getId());
-        createGroupInfo.setModerator(user.getId());
-        groupService.createGroup(createGroupInfo);
+        createGroup();
     }
 
     @Test
@@ -60,6 +58,12 @@ public class GroupServiceTest extends BaseTest {
         assertThat(groups.get(0).getName(), equalTo("Launch 4"));
         assertThat(groups.get(0).getModerator().getUsername(), equalTo(user.getUsername()));
         assertThat(groups.get(0).getTournament().getName(), equalTo(tournament.getName()));
+
+        try {
+            createGroup();
+        } catch (DataInvalidException exception) {
+            assertThat(exception.getMessage(), equalTo("{validation.group.existName.message}"));
+        }
     }
 
     @Test
@@ -107,10 +111,60 @@ public class GroupServiceTest extends BaseTest {
         }
     }
 
+    @Test
+    public void testCheckGroupByRole() {
+        User tester4 = makeUser("Tester4");
+        userRepo.save(tester4);
+        User admin = makeUser("admin");
+        admin.setRole(Role.ADMIN);
+        userRepo.save(admin);
+
+        mockLoginUser(user);
+        List<Group> groups = groupService.getGroupByRole(tournament.getId());
+        assertThat(1, equalTo(groups.size()));
+
+        mockLoginUser(tester4);
+        groups = groupService.getGroupByRole(tournament.getId());
+        assertThat(0, equalTo(groups.size()));
+
+        mockLoginUser(admin);
+        groups = groupService.getGroupByRole(tournament.getId());
+        assertThat(1, equalTo(groups.size()));
+
+    }
+
+    @Test
+    public void testCheckMemberPermission() {
+        User tester5 = makeUser("Tester5");
+        userRepo.save(tester5);
+
+        Group group = new Group();
+        group.setMembers(Arrays.asList(getDefaultUser()));
+        group.setModerator(tester5);
+        group.setName("Launch 4");
+        group.setTournament(tournament);
+        groupRepo.save(group);
+
+        mockLoginUser(tester5);
+        try {
+            groupService.checkMemberPermission(group.getId());
+        } catch (DataInvalidException exception) {
+            assertThat(exception.getMessage(), equalTo("{exception.unauthorized}"));
+        }
+    }
+
     @After
     public void deteleData() {
         groupRepo.deleteAll();
         tournamentRepo.deleteAll();
+    }
+
+    private void createGroup() {
+        CreateGroupInfo createGroupInfo = new CreateGroupInfo();
+        createGroupInfo.setName("Launch 4");
+        createGroupInfo.setTournamentId(tournament.getId());
+        createGroupInfo.setModerator(user.getId());
+        groupService.createGroup(createGroupInfo);
     }
 
 }